… COMPREHENSIVE_GAP_ANALYSIS_REPORT_2025.md
… IMMEDIATE_ACTION_PLAN.md
… PHASE_1_VALIDATION.md
… PHASE_2_COMPLETION_REPORT.md
… PHASE_3_COMPLETION_REPORT.md
… PHASE_4_COMPLETION_REPORT.md
… bmad-agent/free-deep-research/package.json
… bmad-agent/free-deep-research/src-tauri/Cargo.toml
… bmad-agent/free-deep-research/src-tauri/src/commands/advanced_analytics.rs
… bmad-agent/free-deep-research/src-tauri/src/commands/mobile_commands.rs
… bmad-agent/free-deep-research/src-tauri/src/services/ml_engine/inference_engine.rs
… bmad-agent/free-deep-research/src-tauri/src/services/ml_engine/model_training.rs
… bmad-agent/free-deep-research/src-tauri/src/services/ml_engine/pattern_analysis.rs
… bmad-agent/free-deep-research/src-tauri/src/services/ml_engine/recommendation_system.rs
… bmad-agent/free-deep-research/src-tauri/src/services/mobile_platform/mobile_api.rs
… bmad-agent/free-deep-research/src-tauri/src/services/mobile_platform/mod.rs
… bmad-agent/free-deep-research/src-tauri/src/services/monitoring/health_checker.rs
… bmad-agent/free-deep-research/src-tauri/src/services/monitoring/metrics_collector.rs
… bmad-agent/free-deep-research/src-tauri/src/services/output_processor/engine.rs
… bmad-agent/free-deep-research/src-tauri/src/services/output_processor/mod.rs
… bmad-agent/free-deep-research/src-tauri/src/services/research_engine/methodology_don_lim.rs
… bmad-agent/free-deep-research/src-tauri/src/services/research_engine/methodology_hybrid.rs
… bmad-agent/free-deep-research/src-tauri/src/services/research_engine/methodology_nick_scamara.rs
… bmad-agent/free-deep-research/src-tauri/src/services/research_engine/workflow_orchestrator.rs
… bmad-agent/free-deep-research/src-tauri/tests/phase2_verification.rs
… bmad-agent/free-deep-research/src/components/advanced-analytics/AdvancedAnalyticsDashboard.tsx
… bmad-agent/free-deep-research/src/components/bmad-integration/BMadIntegrationDashboard.tsx
… bmad-agent/free-deep-research/src/components/bmad-integration/__tests__/BMadIntegrationDashboard.test.tsx
… docker/backend/Dockerfile
… docker/database/init/01-init.sql